Remove AddDevice screen
